remove scatter float cast

+++ src/olmo_core/nn/blt/local_models.py
@@ -1094,8 +1094,8 @@ class LocalDecoder(nn.Module):
                 depool_out_modulated = depool_out
 
             # skip bos - considered boundary
-            h = (depool_out_modulated[:, :-1] + embeds[:, 1:]).float()
-            h_b = (self.boundary_embedding.weight.unsqueeze(0) + prepool_out).float()
+            h = (depool_out_modulated[:, :-1] + embeds[:, 1:])
+            h_b = (self.boundary_embedding.weight.unsqueeze(0) + prepool_out)
 
             # +1 to keep multiple of
             h_with_b = torch.zeros(
@@ -1133,27 +1133,27 @@ class LocalDecoder(nn.Module):
                 block = self.blocks[str(block_idx)]
                 h_with_b = block(h_with_b)
 
-            h_with_b = h_with_b.float()
+            h_with_b = h_with_b.to(dtype)
 
             h_for_true_boundaries = torch.gather(
                 h_with_b,
                 dim=1,
                 # can't predict first boundary / bos
                 index=(b_indices[:, 1:] - 1).unsqueeze(-1).expand(-1, -1, self.d_model),
-            ).to(dtype)
+            )
             h_for_all_boundaries = torch.gather(
                 h_with_b,
                 dim=1,
                 index=non_b_indices.unsqueeze(-1).expand(-1, -1, self.d_model),
-            ).to(dtype)
+            )
             h_for_logits = torch.gather(
                 h_with_b,
                 dim=1,
                 index=(non_b_indices - 1).unsqueeze(-1).expand(-1, -1, self.d_model),
-            ).to(dtype)
+            )
 
             # [:-1] to strip multiple of
-            return (h_for_true_boundaries, h_for_all_boundaries), h_for_logits, h_with_b[:, :-1].to(dtype)
+            return (h_for_true_boundaries, h_for_all_boundaries), h_for_logits, h_with_b[:, :-1]
 
     def _depool_blt(
         self,
